--- includes/hooks/test/Actions.csv
@@ -1,27 +1,0 @@
-Hook,Reference,Needed?,Description
-activated_plugin,https://developer.wordpress.org/reference/hooks/activated_plugin/,,Fires after a plugin has been activated.
-deactivated_plugin,https://developer.wordpress.org/reference/hooks/deactivated_plugin/,,Fires after a plugin is deactivated.
-authenticate,https://developer.wordpress.org/reference/hooks/authenticate/,,Filters whether a set of user login credentials are valid.
-current_screen,https://developer.wordpress.org/reference/hooks/current_screen/,,Fires after the current screen has been set.
-edit_user_profile_update,https://developer.wordpress.org/reference/hooks/edit_user_profile_update/,,Fires before the page loads on the ‘Edit User’ screen.
-registered_post_type,https://developer.wordpress.org/reference/hooks/registered_post_type/,,Fires after a post type is registered.
-registered_taxonomy,https://developer.wordpress.org/reference/hooks/registered_taxonomy/,,Fires after a taxonomy is registered.
-registered_taxonomy_for_object_type,https://developer.wordpress.org/reference/hooks/registered_taxonomy_for_object_type/,,Fires after a taxonomy is registered for an object type.
-register_form,https://developer.wordpress.org/reference/hooks/register_form/,,Fires following the ‘Email’ field in the user registration form.
-register_meta_args,https://developer.wordpress.org/reference/hooks/register_meta_args/,,Filters the registration arguments when registering meta.
-register_new_user,https://developer.wordpress.org/reference/hooks/register_new_user/,,Fires after a new user registration has been recorded.
-register_post,https://developer.wordpress.org/reference/hooks/register_post/,,"Fires when submitting registration form data, before the user is created."
-remove_user_role,https://developer.wordpress.org/reference/hooks/remove_user_role/,,Fires immediately after a role as been removed from a user.
-render_block,https://developer.wordpress.org/reference/hooks/render_block/,,Filters the content of a single block.
-render_block_data,https://developer.wordpress.org/reference/hooks/render_block_data/,,"Filters the block being rendered in render_block(), before it’s processed."
-plugin_loaded,https://developer.wordpress.org/reference/hooks/plugin_loaded/,,Fires once a single activated plugin has loaded.
-password_reset,https://developer.wordpress.org/reference/hooks/password_reset/,,Fires before the user’s password is reset.
-plugins_loaded,https://developer.wordpress.org/reference/hooks/plugins_loaded/,,Fires once activated plugins have loaded.
-muplugins_loaded,https://developer.wordpress.org/reference/hooks/muplugins_loaded/,,Fires once all must-use and network-activated plugins have loaded.
-mu_plugin_loaded,https://developer.wordpress.org/reference/hooks/mu_plugin_loaded/,,Fires once a single must-use plugin has loaded.
-granted_super_admin,https://developer.wordpress.org/reference/hooks/granted_super_admin/,,Fires after the user is granted Super Admin privileges.
-mature_blog,https://developer.wordpress.org/reference/hooks/mature_blog/,,Fires when the ‘mature’ status is added to a site.
-init,https://developer.wordpress.org/reference/hooks/init/,,Fires after WordPress has finished loading but before any headers are sent.
-load-page.php,https://developer.wordpress.org/reference/hooks/load-page.php/,,
-load-page-new.php,https://developer.wordpress.org/reference/hooks/load-page-new.php/,,
-invite_user,https://developer.wordpress.org/reference/hooks/invite_user/,,
