--- includes/hooks/test/Delete.csv
@@ -1,23 +1,0 @@
-Hook,Reference,Needed?,Description
-delete_post,https://developer.wordpress.org/reference/hooks/delete_post/,,Fires immediately before a post is deleted from the database.
-deleted_post,https://developer.wordpress.org/reference/hooks/deleted_post/,,Fires immediately after a post is deleted from the database.
-after_delete_post,https://developer.wordpress.org/reference/hooks/after_delete_post/,,"Fires after a post is deleted, at the conclusion of wp_delete_post()."
-before_delete_post,https://developer.wordpress.org/reference/hooks/before_delete_post/,,"Fires before a post is deleted, at the start of wp_delete_post()."
-clean_attachment_cache,https://developer.wordpress.org/reference/hooks/clean_attachment_cache/,,Fires after the given attachment’s cache is cleaned.
-clean_comment_cache,https://developer.wordpress.org/reference/hooks/clean_comment_cache/,,Fires immediately after a comment has been removed from the object cache.
-clean_object_term_cache,https://developer.wordpress.org/reference/hooks/clean_object_term_cache/,,Fires after the object term cache has been cleaned.
-clean_page_cache,https://developer.wordpress.org/reference/hooks/clean_page_cache/,,Fires immediately after the given page’s cache is cleaned.
-clean_post_cache,https://developer.wordpress.org/reference/hooks/clean_post_cache/,,Fires immediately after the given post’s cache is cleaned.
-clean_taxonomy_cache,https://developer.wordpress.org/reference/hooks/clean_taxonomy_cache/,,Fires after a taxonomy’s caches have been cleaned.
-clean_term_cache,https://developer.wordpress.org/reference/hooks/clean_term_cache/,,Fires once after each taxonomy’s term cache has been cleaned.
-clean_user_cache,https://developer.wordpress.org/reference/hooks/clean_user_cache/,,Fires immediately after the given user’s cache is cleaned.
-deleted_comment,https://developer.wordpress.org/reference/hooks/deleted_comment/,,Fires immediately after a comment is deleted from the database.
-deleted_link,https://developer.wordpress.org/reference/hooks/deleted_link/,,Fires after a link has been deleted.
-deleted_option,https://developer.wordpress.org/reference/hooks/deleted_option/,,Fires after an option has been deleted.
-deleted_plugin,https://developer.wordpress.org/reference/hooks/deleted_plugin/,,Fires immediately after a plugin deletion attempt.
-deleted_postmeta,https://developer.wordpress.org/reference/hooks/deleted_postmeta/,,Fires immediately after deleting metadata for a post.
-deleted_term_relationships,https://developer.wordpress.org/reference/hooks/deleted_term_relationships/,,Fires immediately after an object-term relationship is deleted.
-deleted_term_taxonomy,https://developer.wordpress.org/reference/hooks/deleted_term_taxonomy/,,Fires immediately after a term taxonomy ID is deleted.
-deleted_transient,https://developer.wordpress.org/reference/hooks/deleted_transient/,,Fires after a transient is deleted.
-deleted_user,https://developer.wordpress.org/reference/hooks/deleted_user/,,Fires immediately after a user is deleted from the database.
-deleted_usermeta,https://developer.wordpress.org/reference/hooks/deleted_usermeta/,,
