--- includes/hooks/test/Edit_Modify.csv
@@ -1,15 +1,0 @@
-Hook,Reference,Needed?,Description
-edit_post,https://developer.wordpress.org/reference/hooks/edit_post/,✔,Fires once an existing post has been updated.
-save_post,https://developer.wordpress.org/reference/hooks/save_post/,,Fires once a post has been saved.
-post_updated,https://developer.wordpress.org/reference/hooks/post_updated/,,Fires once an existing post has been updated.
-trashed_post,https://developer.wordpress.org/reference/hooks/trashed_post/,,Fires after a post is sent to the trash.
-untrash_post,https://developer.wordpress.org/reference/hooks/untrash_post/,,Fires before a post is restored from the trash.
-after_password_reset,https://developer.wordpress.org/reference/hooks/after_password_reset/,,Fires after the user’s password is reset.
-after_switch_theme,https://developer.wordpress.org/reference/hooks/after_switch_theme/,,Fires on the first WP load after a theme switch if the old theme still exists.
-attachment_updated,https://developer.wordpress.org/reference/hooks/attachment_updated/,,Fires once an existing attachment has been updated.
-customize_save_after,https://developer.wordpress.org/reference/hooks/customize_save_after/,,Fires after Customize settings have been saved.
-customize_save,https://developer.wordpress.org/reference/hooks/customize_save/,,"Fires once the theme has switched in the Customizer, but before settings have been saved."
-edited_term,https://developer.wordpress.org/reference/hooks/edited_term/,,"Fires after a term has been updated, and the term cache has been cleaned."
-edited_term_taxonomies,https://developer.wordpress.org/reference/hooks/edited_term_taxonomies/,,Fires immediately after a term to delete’s children are reassigned a parent.
-profile_update,https://developer.wordpress.org/reference/hooks/profile_update/,,Fires immediately after an existing user is updated.
-profile_personal_options,https://developer.wordpress.org/reference/hooks/profile_personal_options/,,Fires after the ‘Personal Options’ settings table on the ‘Your Profile’ editing screen.
