+++ src/bootstrap/class-jettison-deactivator.php
