+++ src/bootstrap/class-jettison-uninstaller.php
